+++ Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic/Exceptions/UserIsAlreadyRegisteredException.cs
@@ -2,6 +2,6 @@ using Texnokaktus.ProgOlymp.Common.Contracts.Grpc.YandexContest;
 
 namespace Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic.Exceptions;
 
-public class UserIsAlreadyRegisteredException(int contestStageId, string yandexIdLogin)
+public class UserIsAlreadyRegisteredException(long contestStageId, string yandexIdLogin)
     : RpcApplicationException(ErrorType.UserIsAlreadyRegistered,
                               $"User {yandexIdLogin} is already registered for Contest Stage {contestStageId}");
