+++ Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic/Exceptions/UserIsNotRegisteredException.cs
@@ -2,6 +2,6 @@
 
 namespace Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic.Exceptions;
 
-public class UserIsNotRegisteredException(int contestStageId, string yandexIdLogin)
+public class UserIsNotRegisteredException(long contestStageId, string yandexIdLogin)
     : RpcApplicationException(ErrorType.UserIsNotRegistered,
                               $"User {yandexIdLogin} is not registered for Contest Stage {contestStageId}");
