+++ Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic/Services/ParticipantService.cs
@@ -5,18 +5,18 @@ namespace Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic.Services;
 
 internal class ParticipantService(IUnitOfWork unitOfWork) : IParticipantService
 {
-    public async Task<long?> GetContestUserIdAsync(int contestStageId, string participantLogin)
+    public async Task<long?> GetContestUserIdAsync(long contestStageId, string participantLogin)
     {
         var contestUser = await unitOfWork.ContestUserRepository.GetAsync(contestStageId, participantLogin);
         return contestUser?.ContestUserId;
     }
 
-    public async Task AddContestParticipantAsync(int contestStageId, string yandexIdLogin, long contestUserId)
+    public async Task AddContestParticipantAsync(long contestStageId, string yandexIdLogin, long contestUserId)
     {
         unitOfWork.ContestUserRepository.Add(new(contestStageId, yandexIdLogin, contestUserId));
         await unitOfWork.SaveChangesAsync();
     }
 
-    public Task DeleteContestParticipantAsync(int contestStageId, string yandexIdLogin) =>
+    public Task DeleteContestParticipantAsync(long contestStageId, string yandexIdLogin) =>
         unitOfWork.ContestUserRepository.DeleteAsync(new(contestStageId, yandexIdLogin));
 }
