+++ Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic/Services/ParticipantServiceCachingDecorator.cs
@@ -5,22 +5,22 @@ namespace Texnokaktus.ProgOlymp.YandexContestIntegrationService.Logic.Services;
 
 internal class ParticipantServiceCachingDecorator(IParticipantService service, IMemoryCache cache) : IParticipantService
 {
-    public Task<long?> GetContestUserIdAsync(int contestStageId, string participantLogin) =>
+    public Task<long?> GetContestUserIdAsync(long contestStageId, string participantLogin) =>
         cache.GetOrCreateAsync(GetCacheKey(contestStageId, participantLogin),
                                _ => service.GetContestUserIdAsync(contestStageId, participantLogin));
 
-    public async Task AddContestParticipantAsync(int contestStageId, string yandexIdLogin, long contestUserId)
+    public async Task AddContestParticipantAsync(long contestStageId, string yandexIdLogin, long contestUserId)
     {
         await service.AddContestParticipantAsync(contestStageId, yandexIdLogin, contestUserId);
         cache.Remove(GetCacheKey(contestStageId, yandexIdLogin));
     }
 
-    public async Task DeleteContestParticipantAsync(int contestStageId, string yandexIdLogin)
+    public async Task DeleteContestParticipantAsync(long contestStageId, string yandexIdLogin)
     {
         await service.DeleteContestParticipantAsync(contestStageId, yandexIdLogin);
         cache.Remove(GetCacheKey(contestStageId, yandexIdLogin));
     }
 
-    private static string GetCacheKey(int contestStageId, string participantLogin) =>
+    private static string GetCacheKey(long contestStageId, string participantLogin) =>
         $"Participant:{contestStageId}:{participantLogin}";
 }
